requirements.txt
# Core framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2

# Environment
python-dotenv==1.0.1

# Emotion detection
transformers==4.44.2
torch==2.4.0
tokenizers==0.19.1

# TTS — Primary
elevenlabs==1.9.0

# TTS — Fallback (offline)
pyttsx3==2.90

# Utilities
httpx==0.27.0
aiofiles==24.1.0
